Add missing xmldoc, rename StaticFieldStorage to StaticFields. Move to Runtime namespace.
